refactor: split __init__.py
This is the first step before replacing argparse by cli
