+++ mergify_cli/stack/edit.py
@@ -1,0 +1,11 @@
+import argparse
+import os
+
+from mergify_cli import utils
+
+
+async def stack_edit(_: argparse.Namespace) -> None:
+    os.chdir(await utils.git("rev-parse", "--show-toplevel"))
+    trunk = await utils.get_trunk()
+    base = await utils.git("merge-base", trunk, "HEAD")
+    os.execvp("git", ("git", "rebase", "-i", f"{base}^"))  # noqa: S606
