+++ mergify_cli/tests/test_mergify_cli.py
@@ -22,7 +22,9 @@ from unittest import mock
 import pytest
 import respx
 
-import mergify_cli
+from mergify_cli import cli
+from mergify_cli import utils
+from mergify_cli.stack import push
 from mergify_cli.tests import utils as test_utils
 
 
@@ -79,14 +81,14 @@ def git_mock(
         output="",
     )
 
-    with mock.patch("mergify_cli.git", git_mock_object):
+    with mock.patch("mergify_cli.utils.git", git_mock_object):
         yield git_mock_object
 
 
 @pytest.mark.usefixtures("_git_repo")
 async def test_cli_help(capsys: pytest.CaptureFixture[str]) -> None:
     with pytest.raises(SystemExit, match="0"):
-        await mergify_cli.parse_args(["--help"])
+        await cli.parse_args(["--help"])
 
     stdout = capsys.readouterr().out
     assert "usage: " in stdout
@@ -96,22 +98,22 @@ async def test_cli_help(capsys: pytest.CaptureFixture[str]) -> None:
 
 @pytest.mark.usefixtures("_git_repo")
 async def test_get_branch_name() -> None:
-    assert await mergify_cli.git_get_branch_name() == "main"
+    assert await utils.git_get_branch_name() == "main"
 
 
 @pytest.mark.usefixtures("_git_repo")
 async def test_get_target_branch() -> None:
-    assert await mergify_cli.git_get_target_branch("main") == "main"
+    assert await utils.git_get_target_branch("main") == "main"
 
 
 @pytest.mark.usefixtures("_git_repo")
 async def test_get_target_remote() -> None:
-    assert await mergify_cli.git_get_target_remote("main") == "origin"
+    assert await utils.git_get_target_remote("main") == "origin"
 
 
 @pytest.mark.usefixtures("_git_repo")
 async def test_get_trunk() -> None:
-    assert await mergify_cli.get_trunk() == "origin/main"
+    assert await utils.get_trunk() == "origin/main"
 
 
 @pytest.mark.parametrize(
@@ -124,7 +126,7 @@ async def test_get_trunk() -> None:
 )
 def test_check_local_branch_valid(valid_branch_name: str) -> None:
     # Should not raise an error
-    mergify_cli.check_local_branch(
+    push.check_local_branch(
         branch_name=valid_branch_name,
         branch_prefix="prefix",
     )
@@ -132,10 +134,10 @@ def test_check_local_branch_valid(valid_branch_name: str) -> None:
 
 def test_check_local_branch_invalid() -> None:
     with pytest.raises(
-        mergify_cli.LocalBranchInvalidError,
+        push.LocalBranchInvalidError,
         match="Local branch is a branch generated by Mergify CLI",
     ):
-        mergify_cli.check_local_branch(
+        push.check_local_branch(
             branch_name="prefix/my-branch/I29617d37762fd69809c255d7e7073cb11f8fbf50",
             branch_prefix="prefix",
         )
@@ -212,7 +214,7 @@ async def test_stack_create(
         200,
     )
 
-    await mergify_cli.stack_push(
+    await push.stack_push(
         github_server="https://api.github.com/",
         token="",
         skip_rebase=False,
@@ -304,7 +306,7 @@ async def test_stack_create_single_pull(
     )
     respx_mock.get("/repos/user/repo/issues/1/comments").respond(200, json=[])
 
-    await mergify_cli.stack_push(
+    await push.stack_push(
         github_server="https://api.github.com/",
         token="",
         skip_rebase=False,
@@ -394,7 +396,7 @@ async def test_stack_update_no_rebase(
     )
     respx_mock.patch("/repos/user/repo/issues/comments/456").respond(200)
 
-    await mergify_cli.stack_push(
+    await push.stack_push(
         github_server="https://api.github.com/",
         token="",
         skip_rebase=True,
@@ -484,7 +486,7 @@ async def test_stack_update(
     )
     respx_mock.patch("/repos/user/repo/issues/comments/456").respond(200)
 
-    await mergify_cli.stack_push(
+    await push.stack_push(
         github_server="https://api.github.com/",
         token="",
         skip_rebase=False,
@@ -573,7 +575,7 @@ async def test_stack_update_keep_title_and_body(
     )
     respx_mock.patch("/repos/user/repo/issues/comments/456").respond(200)
 
-    await mergify_cli.stack_push(
+    await push.stack_push(
         github_server="https://api.github.com/",
         token="",
         skip_rebase=False,
@@ -602,7 +604,7 @@ async def test_stack_on_destination_branch_raises_an_error(
     git_mock.mock("rev-parse", "--abbrev-ref", "HEAD", output="main")
 
     with pytest.raises(SystemExit, match="1"):
-        await mergify_cli.stack_push(
+        await push.stack_push(
             github_server="https://api.github.com/",
             token="",
             skip_rebase=False,
@@ -622,7 +624,7 @@ async def test_stack_without_common_commit_raises_an_error(
     git_mock.mock("merge-base", "--fork-point", "origin/main", output="")
 
     with pytest.raises(SystemExit, match="1"):
-        await mergify_cli.stack_push(
+        await push.stack_push(
             github_server="https://api.github.com/",
             token="",
             skip_rebase=False,
@@ -636,9 +638,9 @@ async def test_stack_without_common_commit_raises_an_error(
 @pytest.mark.parametrize(
     ("default_arg_fct", "config_get_result", "expected_default"),
     [
-        (mergify_cli.get_default_keep_pr_title_body, "true", True),
+        (cli.get_default_keep_pr_title_body, "true", True),
         (
-            lambda: mergify_cli.get_default_branch_prefix("author"),
+            lambda: utils.get_default_branch_prefix("author"),
             "dummy-prefix",
             "dummy-prefix",
         ),
@@ -652,7 +654,7 @@ async def test_defaults_config_args_set(
     config_get_result: bytes,
     expected_default: bool,
 ) -> None:
-    with mock.patch.object(mergify_cli, "_run_command", return_value=config_get_result):
+    with mock.patch.object(utils, "run_command", return_value=config_get_result):
         assert (await default_arg_fct()) == expected_default
 
 
@@ -669,7 +671,7 @@ async def test_default(
     args: list[str],
 ) -> None:
     git_mock.default_cli_args()
-    parsed = await mergify_cli.parse_args(args)
+    parsed = await cli.parse_args(args)
     assert parsed.action == "stack"
     assert parsed.stack_action == "push"
     assert parsed.skip_rebase
@@ -679,6 +681,6 @@ async def test_parse_edit(
     git_mock: test_utils.GitMock,
 ) -> None:
     git_mock.default_cli_args()
-    parsed = await mergify_cli.parse_args(["stack", "edit"])
+    parsed = await cli.parse_args(["stack", "edit"])
     assert parsed.action == "stack"
     assert parsed.stack_action == "edit"
